pyproject.toml
[project]
name = "basic-reference-rag-implementation"
version = "0.1.0"
description = "AI agent service built with LangGraph, FastAPI and Chainlit"
readme = "README.md"
authors = [{ name = "Arian Pasquali", email = "arianpasquali@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.11"

dependencies = [
    # Core LangChain & LangGraph
    "langchain-core~=0.3.33",
    "langchain-community ~=0.3.16",
    "langchain-chroma ~=0.2.3",
    "langchain-openai~=0.3.0",
    "langchain-text-splitters ~=0.3.0",
    "langgraph~=0.6.6",
    # Vector Databases
    "chromadb >=1.1.0",
    # UI & Web Framework
    "chainlit >=2.6.3",
    # Data Processing
    "pandas ~=2.2.3",
    "pypdf ~=5.3.0",
    # Configuration & Validation
    "pydantic ~=2.10.1",
    "pydantic-settings ~=2.6.1",
    "python-dotenv ~=1.0.0",
    # Observability
    "orq-ai-sdk",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    # Utilities
    "tiktoken >=0.8.0",
]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

eval = [
    "evaluatorq",
    "ragas ~=0.1.0",
    "pytest-benchmark",
    "tabulate",
    "matplotlib",
    "seaborn",
    "pandas",
    "numpy",
]

client = [
    "pydantic ~=2.10.1",
    "python-dotenv ~=1.0.0",
    "chainlit >=2.6.3",
]

[tool.ruff]
line-length = 100
target-version = "py39"  # Changed to match Python 3.9 compatibility
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".pytest_cache",
    ".mypy_cache",
    ".ruff_cache",
    "docs/",
    "lancedb/",
    "scripts/__pycache__/",
]

[tool.ruff.lint]
# Enable core rule sets for practical development
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
]

ignore = [
    "E501",   # Line too long (let ruff format handle this)
    "W293",   # Blank line contains whitespace
    "W291",   # Trailing whitespace
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "lancedb",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S106", "S108"]  # Allow assert, hardcoded passwords in tests
"scripts/**/*.py" = ["S101", "S603", "S607"]  # Allow scripts to be more permissive

[tool.ruff.lint.isort]
known-first-party = ["assistant", "shared"]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Security settings for bandit integration
[tool.bandit]
exclude_dirs = ["tests", "scripts/test_*"]
skips = ["B101", "B601"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"

[tool.pytest_env]
OPENAI_API_KEY = "sk-fake-openai-key"

[tool.mypy]
plugins = "pydantic.mypy"
